lab1/wine/huggingface-simulation/requirements.txt
gradio==4.4.1
hopsworks==3.4.3
joblib==1.3.2
pandas==2.0.3
Pillow==10.1.0
Requests==2.31.0
scikit-learn==1.3.2
